+++ my-video-room/css/visualiser-override.css
@@ -1,5 +1,3 @@
-
-
 .nav-tab-wrapper{
 display:none
 }
