+++ my-video-room/views/visualiser/admin-settings-roombuilder.php
@@ -26,7 +26,7 @@ return function (
 	array $messages = array()
 ): string {
 
-	wp_enqueue_script( 'frametab' );
+	wp_enqueue_script( 'mvr-frametab' );
 	wp_enqueue_style( 'visualiser' );
 	$render = require __DIR__ . '/header.php';
 	// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- Already escaped.
